AC-880 CR Addressed code review issue where any file could be deleted by the UnixRouteListener when unlink_existing was true

+++ src/com/crankuptheamps/ttunnel/UnixRouteListener.java
@@ -6,6 +6,7 @@ import java.net.UnixDomainSocketAddress;
 import java.nio.channels.ServerSocketChannel;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.attribute.BasicFileAttributes;
 
 public class UnixRouteListener implements RouteListener {
 
@@ -19,7 +20,17 @@ public class UnixRouteListener implements RouteListener {
         }
         if (Files.exists(path)) {
             if (endpoint.isUnlinkExisting()) {
-                Files.delete(path);
+                try {
+                    BasicFileAttributes attrs = Files.readAttributes(path, BasicFileAttributes.class);
+                    // If the file is an existing Unix domain socket file, delete it
+                    if (attrs.isOther() && !Files.isSymbolicLink(path)) {
+                        Files.delete(path);
+                    } else {
+                        throw new IOException("Cannot delete file since is not a Unix domain socket file.");
+                    }
+                } catch (Exception e) {
+                    throw new IOException("Cannot delete file since it is not a Unix domain socket file.", e);
+                }
             } else {
                 throw new IOException("Unix domain socket path already exists: " + path);
             }
@@ -40,3 +51,4 @@ public class UnixRouteListener implements RouteListener {
         }
     }
 }
+
